Diakonov_Denys/back/requirements.txt
django < 4.0
djangorestframework
markdown
django-filter
psycopg2
drf-yasg
django-cors-headers
docxtpl
django-heroku
gunicorn